requirements.txt
streamlit
pandas
plotly
numpy
prettymapp
osmnx
matplotlib
scikit-learn
contextily
geopandas
shapely
pillow
requests